requirements.txt
# Core dependencies for VLM PDF parser
python-dotenv>=1.0.0
openai>=1.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
requests>=2.31.0
PyMuPDF>=1.24.7
boto3>=1.34.0
docling
instructor[bedrock]